requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
passlib[bcrypt]
python-jose
redis